requirements.txt
numpy
pillow
bokeh
scikit-learn
umap-learn